add correct file names and extensions to report routes
Also, report routes for PDF files no longer send the file as an
attachment. Downloading the file is still possible, but handling now
depends on the browser. The upside is that in cases where we waiting on
the creation of the PDF, the download does not start, and the user
remains on the page stating "The report is still being generated...".